+++ src/components/options/ColumnAliasesEditor.tsx
