+++ src/components/options/ColumnWidthHintsEditor.tsx
